+++ src/game/units/Vehicle.cc
@@ -35,7 +35,7 @@ void Vehicle::dropOff(size_t crew_index, const Coords& location)
   map->stashPopUnit(*unit);
 
   _dropped.push_back(_crew[crew_index]);
-  _crew.erase(_crew.begin() + crew_index);
+  _crew.erase(_crew.begin() + static_cast<int> (crew_index));
 }
